完成发布问题编辑功能

+++ src/main/java/com/zxj/community/controller/QuestionController.java
@@ -20,12 +20,11 @@ public class QuestionController {
     private QuestionService questionService;
 
     @GetMapping("/question/{id}")
-    public String question(@PathVariable(name = "id")Integer id,
+    public String question(@PathVariable(name = "id")Long id,
                            Model model){
 
         QuestionDTO questionDTO = questionService.getById(id);
         model.addAttribute("question", questionDTO);
         return "question";
-
     }
 }
